refactor: domain으로 PrizeResults
